+++ packages/esbuild-pilet-plugin/src/importmap-plugin.ts
@@ -1,5 +1,5 @@
-import type { SharedDependency } from "piral-cli";
-import type { PluginObj } from "@babel/core";
+import type { SharedDependency } from 'piral-cli';
+import type { PluginObj } from '@babel/core';
 
 export interface PluginOptions {
   importmap: Array<SharedDependency>;
@@ -13,14 +13,12 @@ export default function babelPlugin({ types }): PluginObj {
 
         path.traverse({
           Literal(path) {
-            if (path.node.type === "StringLiteral") {
+            if (path.node.type === 'StringLiteral') {
               const name = path.node.value;
               const entry = importmap.find((m) => m.name === name);
 
               if (entry) {
-                path.replaceWith(
-                  types.stringLiteral((entry as any).requireId || entry.id)
-                );
+                path.replaceWith(types.stringLiteral((entry as any).requireId || entry.id));
               }
             }
           },