+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## 0.15.0 (tbd)
 
 - Moved package to dedicated repository
+- Added `esbuild-codegen-plugin`
+- Added `esbuild-pilet-plugin`